+++ ElvUI_LuckyoneUI/Init.lua
@@ -46,12 +46,6 @@ local toons = {
 	["Luckydk-Everlook"]			= true, -- [Alliance] DK
 	["Luckykek-Everlook"]			= true, -- [Alliance] Shaman
 	["Luckyone-Giantstalker"]		= true, -- [Alliance] Paladin
-	-- Testing
-	["Luckyone-Koranos"]			= true, -- [Beta] Druid
-	["Kekleox-Tyrhold"]				= true, -- [Beta] Evoker
-	["Luckyevoker-Valdrakken"]		= true, -- [Beta] Evoker
-	["Luckyone-Valdrakken"]			= true, -- [Beta] Druid
-	["Luckyone-Nobundo"]			= true, -- [PTR] Druid
 }
 
 -- Constants
@@ -62,7 +56,7 @@ L1UI.DefaultTexture = 'Minimalist'
 L1UI.Logo = 'Interface\\AddOns\\ElvUI_LuckyoneUI\\Media\\Textures\\Clover.tga'
 L1UI.Me = toons[name]
 L1UI.Name = '|cff4beb2cLuckyoneUI|r'
-L1UI.RequiredVersion = 13.02
+L1UI.RequiredVersion = 13.06
 L1UI.Version = GetAddOnMetadata(addon, 'Version')
 
 function L1UI:Initialize()